Make some SVG improvements
The sed arg list too long error still happens for several files, though.

+++ svg-templates.sh
@@ -1,9 +1,10 @@
-#!/bin/sh
+#!/bin/bash
 
 # Replace {{ mustache }} syntax in index.html with corresponding SVGs
-find src/_svgs -type f -name "*.svg" | while read -r file; do
+find ./src/_svgs -type f -name "*.svg" | while read -r file; do
   name="$(basename "$file" | cut -d . -f 1)"
   contents="$(cat "$file")"
-  sed -i.bak "s;{{ $name }};$contents;g" public/index.html
-  rm public/index.html.bak
+  echo "$name"
+  sed -i.bak "s;{{ $name }};$contents;g" ./public/index.html
+  if [ -f ./public/index.html.bak ]; then rm public/index.html.bak; fi
 done
